+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 * Add checks for multiple okapi interfaces on user's details screen. Fixes UIU-1600.
 * Display automated patron blocks on User Information page. Refs UIU-1273.
 * On user-edit screen, show "send reset password link" whenever username is present. Refs UIU-1672.
+* Display automated patron blocks on renewing in loans context. Refs UIU-1276.
 
 ## [3.0.0](https://github.com/folio-org/ui-users/tree/v3.0.0) (2020-03-17)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.26.0...v3.0.0)